[v8][api] Update chrome.runtime.lastError implementation
... to use v8::Object::SetNativeDataProperty() instead of soon-to-be-
deprecated v8::Object::SetAccessor().

This CL slightly changes the behaviour in the following scenario and
makes it work according to the JavaScript spec which wasn't the case
before.
Imagine the following setup:
```
  var proto = chrome.runtime;
  var obj = Object.create(proto);
  obj.lastError = 42; // [*]
```
Before this CL `[*]` was a no-op while after it'll define "lastError"
property on `obj`.

Given that
 - all the runtime.lastError is supposed to be set by Chromium and
   and read by a particular Chrome extension implementation,
 - we can not think of a meaningful usage scenario relying on setting
   those functions via prototype chain,
I think this change is acceptable.

Bug: 336325111
Change-Id: Idef66fdfafd80f9fe25a3ff7383b6ec727441744
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5487600
Reviewed-by: Tim <tjudkins@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293563}