language_packs: Parameterise font tests over loading during login
This is to ensure that code which changes the behaviour of the font
service with the "load fonts after download during login" feature param
does not accidentally affect other behaviour.

As no code currently uses the feature param, this parameterisation
should be a no-op.

Bug: b:331667343
Change-Id: Ibd38694787f025d3ba33fcf7c8e53d8e249c8b7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5479289
Reviewed-by: Curtis McMullan <curtismcmullan@chromium.org>
Commit-Queue: Michael Cui <mlcui@google.com>
Cr-Commit-Position: refs/heads/main@{#1293565}