Fix CHECK crash in autofill popup.
When mouse is used to select the control part of
a suggestion and then with keyboard (UP or DOWN)
the focus (selected suggestion) moves to a suggestion
without control (e.g. a footer suggestion), the CHECK
before scheduling SetRowWithOpenSubPopup call in
SetSelectedCell raises an error, as the sub-popup
cannot be open on a suggestion without children.
This CL fixes it by making sure that when the newly
selected suggestion doesn't have children the selected
cell gets the `CellType::kContent` type.

Bug: 40923811
Change-Id: I27166ad002b5cfad330a1d544cd38669a6c1012c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5494070
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Dmitry Vykochko <vykochko@google.com>
Cr-Commit-Position: refs/heads/main@{#1293568}