[FedCM] Fix button flow crash
There was a recent change to reset account_selection_view more often:
https://crrev.com/c/5479590 for multi IDP. Unfortunately, it crashes
button flows and we didn't have tests to catch that. This patch makes it
such that we only reset the account_selection_view in the multi IDP
case and in the case where we switch from a modal to a type of dialog
which does not have modal support.

Bug: 337075386
Change-Id: I52233037a49f760f8bb1d48d370881d284d8d365
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5496132
Commit-Queue: Yi Gu <yigu@chromium.org>
Reviewed-by: Yi Gu <yigu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293628}