[DBSC] Add support for a new header format for session registrations
The current format for "Sec-Session-Google-Registration" has an
important flaw - it doesn't support multiple registrations in the same
response.

This CL adds support for the new header format, matching
"Sec-Session-Registration" from
https://github.com/WICG/dbsc?tab=readme-ov-file#start-session. We are
not using the standard header name for the google.com prototype impl.

The standard format is not finalized yet so the new header is guarded
behind a feature flag for now.

The code keeps supporting the old registration header but the new
header takes priority over the old one: if both headers are present in
a response, only the new format is taken into consideration. This is
needed to support a transition period during which the server will send
headers in both formats for keeping compatibility with old clients.

The current impl doesn't support multiple sessions yet, so registrations
are still guarded by
`switches::kEnableBoundSessionCredentialsExclusiveRegistrationPath`.

The unit tests are updated to cover both header formats and some unique
tests for the new header are added.

Bug: b/333731555
Change-Id: Ie884d5d8494b715125d3186c563c63d24bc56e39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5444639
Reviewed-by: Monica Basta <msalama@chromium.org>
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Kristian Monsen <kristianm@chromium.org>
Auto-Submit: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293632}