Use LINT.IfChange for sync UMA enums
When the enum name in C++/Java diverges from the one in XML, the
IFTTT label is always the XML name (because this was easier to
automate).

NO_IFTTT=There seems to be a bug when setting up new IFTTTS, but next changes warn correctly, e.g. crrev.com/c/5483448

Change-Id: I8834715b5ee1645f1d9cc20a78ddf1b4c5d9ec4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5480135
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/main@{#1293671}