[ios] Fix scroll position when using optimised session restoration
When using WKWebView's native session data, the scroll position is
saved. This causes the view to be incorrectly positioned when user
scroll the view, quit the app and reload it, if the native session
data is used and kSmoothScrollingDefault is enabled.

This was only visible with optimised session restoration since the
legacy session restoration had a regression causing it to fail to
load the native session data (thus hiding the regression from the
fullscreen code).

This CL change SessionRestorationServiceImpl to not load the native
session data if kSmoothScrollingDefault is enabled, behaving in the
same way as LegacySessionRestorationService (bug included).

    Fixing LegacySessionRestorationService to correctly load the
    native session would make this bug happen for all users. This
    is undesirable.

Bug: 337816888
Fixed: 336613927
Change-Id: Ifb3f448cf30dd6f2c8044dd4cfd707a2ffec590c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5490735
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293694}