[iOS] Enable Bookmarks/RL when sgining in from the type manager
Currently, after signing in from the promo in the bookmarks or reading
list manager with the corresponding type disabled, another promo shows
that opens account settings to allow the user to enable the type.

To avoid double-promo flow, this CL turns ON the corresponding data type
on signing in from the data specific access point.
And handles disabling the type if the user taps "undo" button from the
sign-in confirmation snackbar.

This CL:
- Introduces new post sign-in actions for enabling the types
- Sets these actions depending on the sign-in access points
- Passes the post sign-in actions to the performer
- Updates EG tests to match the new one-promo flow.

Fixed: 41488185
Change-Id: Iab34b3cf2aec928716a1f993e0f173c264314b37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5489141
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Jood Hajeer <jood@google.com>
Cr-Commit-Position: refs/heads/main@{#1293695}