Set request's referrer policy, not just string, for window.open
https://html.spec.whatwg.org/multipage/nav-history-apis.html#window-open-steps:~:text=If%20noreferrer%20is%20true%2C%20then%20set%20noopener%20to%20true%20and%20set%20referrerPolicy%20to%20%22no%2Dreferrer%22

Prior to this change, the referrer string itself was redacted properly,
but the policy was not correctly set to `no-referrer`, meaning that any
downstream users could accidentally use the wrong value.

This change relies on SecurityPolicy::GenerateReferrer to perform the
redaction given the appropriate policy, which is more intuitive and
more closely matches the spec.

Change-Id: If4f855ae02547d96f89de720494e6e521925e373
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5468180
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293732}