[ios][web] Prevent crash in WKHTTPSystemCookieStore at shutdown
WKHTTPSystemCookieStore was using web::Get{UI,IO}ThreadTaskRunner()
to post tasks between UI and IO threads. Using those methods are
only valid when the threads exists and the task executor installed.

This means that calling those methods is invalid while the threads
are being shutdown. As the IO thread is terminated before the UI
thread, there is a short window of time where IO post a task to UI,
the thread IO thread is destroyed and the task tries to post to IO
thread the result.

    This was found writing ChromeBrowserStateManagerImpl tests
    as the test would create a real IO thread, create a real
    ChromeBrowserStateImpl (thus a WKHTTPSystemCookieStore). Then
    the test would terminate the IO then UI thread. The race was
    found as the IO thread sometimes managed to post a task to UI
    thread and the response would cause the test to fail when it
    tried to post the task to IO thread.

Instead change WKHTTPSystemCookieStore to capture a scoped_refptr
to both UI and IO threads' TaskRunner when it is constructed. The
TaskRunner references are valid even after the thread is terminated
(they drop the task when calling PostTask after thread destruction).

To make it easier to work with the TaskRunner without capturing
C++ pointer by accident, use base::OnceCallback<...> as much as
possible in the implementation (using helpers such as BindPostTask).

Bug: 333865629
Change-Id: I5babd7c04c45eb27cc6da8547567ee7814f79301
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5471831
Reviewed-by: Federica Germinario <fedegermi@google.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293733}