Remove chili@chromium.org from chrome/browser/xsurface/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/chili@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ic7a889ff3e37ec69cf3db7f0780e4184180e0a03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5452206
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Ian Wells <iwells@chromium.org>
Reviewed-by: Ian Wells <iwells@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293744}