DrawImageToneMapUtil: Disable tone mapping when source is destination
When uploading an HLG video frame to a WebGL texture, the frame can
sometimes go through VideoImageGenerator to be uploaded, and then
be copied to a HLG texture using a drawImage call.

The call to drawImage will apply tone mapping, even though the goal of
copying HLG to HLG is to do no conversion. As a (presumably temporary)
workaround, disable tone mapping when the source and destination color
spaces are the same.

This is an area that is not well defined by any specification yet,
and will likely need revision in the future.

Bug: 337538021
Change-Id: I502bc29f57d9a00356ffc556b0870f38fe901f0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5490430
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293750}