[Hub] Disable browser controls while visible
On LFF there is a seemingly few pixel "overlap" between the android
toolbar and the Hub's toolbar (at the bottom). For some reason in this
region click events on the Hub can "pass through" to the Toolbar. (This
is not an issue on phones). I suspect this is related to the prior
ability for the toolbar to reparent atop the Tab Switcher (although I
cannot find where this happens). As an interim solution hide the toolbar
Android view while the Hub is fully visible to prevent touch inputs from
passing through to the toolbar.

Fixed: 337616153
Change-Id: I8ddcda80145811774f7e48784960a54dc35d443b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5493335
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293763}