[lensoverlay] Fix text disappearing when issuing a multimodal query.
This moves responsibility for clearing the searchbox thumnail and text
from the searchbox to the overlay controller, which has more information
about the state of the user interactions. The controller now clears the
searchbox text when a thumbnail-creating action is issued and clears the
thumbnail when text is selected in the overlay.

Additionally, this removes last_search_box_text_. This was used to
preserve the latest searchbox query and use it along with a new image
selection to issue a multimodal query. But my understanding is that in the cases where image selection is triggering a query, that should clear the text and simply issue an image query (assuming we want to match the behavior elsewhere).

Bug: b:337290757
Change-Id: I9072d29b10d5633eafb567cd2ea3272e7a74fa81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5494122
Reviewed-by: Duncan Mercer <mercerd@google.com>
Commit-Queue: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293770}