CBD Timeframe Experiment: update tests to work w/ and w/o experiment
In preparation of the field trial testing config for the experiment,
which will run all relevant CBD dialog test suites (also regular non-
experiment ones) without and with the experiment, upgrade the regular
non-experiment test suites to conditionally also work with the
experiment enabled. This is necessary as we don't enable this
experiment for all test suites by default - which is what we'd usually
do for features that will roll out to default at some point (which we
don't here, as this is just an experiment).

The test upgrades boil down to:
1. Switching out regular prefs with experiment prefs in case the
  experiment is on.
2. Selecting a time period before confirming the deletion also in
   regular tests.

Bug: 40283307
Change-Id: I943856063b3db02f67bd1c74842312154a026afe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5490787
Auto-Submit: Rainhard Findling <rainhard@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Commit-Queue: Martin Šrámek <msramek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293793}