Avoid dangling pointer to Textfield::controller_ in c/b/u/v/payments.
This CL is another step towards being able to mark
Textfield::controller_ as not dangling, since this is one of the sites
in which Textfield::controller_ can dangle.

EditorViewController registers itself as the controller for three
(Validating)Textfields, but can be destroyed prior to destruction of
the Textfields. This CL unsets the controller of these Textfields
on destruction of EditorViewController to avoid this from happening.

However, there are also calls paths (e.g. when UpdateContentView() of
the parent class is called, or when
PaymentRequestDialogView::OnDialogClosed() is called) in which the views
are destroyed before the controller is. To also deal with these cases,
the CL extends (the already existing) Stop method to reset all pointers
to views.

Bug: 40277363
Change-Id: I4166ffdc78ac0ccac64fe3cdcbbaa4a319892741
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4926269
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1293802}