Grow HoverButton's height when necessary.
HoverButton does not use LabelButton's Label as content in some cases
and the Label may be multi-line, but it does not overload
GetHeightForWidth, and adjust_height_for_width is false. This causes
incorrect behavior when the label has multiple lines. At the same time,
CalculatePreferredSize(SizeBounds) also needs to be reloaded because
LayoutManager has a cache. This is due to crbugs.com/41357453 which
causes the layout manager to have a cache during the first layout.
Therefore, the cache cannot be invalidated on the second layout. We need
to use GetPreferredSize(SizeBounds) to avoid using cache.

Of course a better approach would be to just overload
CalculatePreferredSize(SizeBounds). But that migration is still in
progress.

Bug: 337075375
Change-Id: Ic8e91b0627a805b319f5a93982fe42670c326838
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5490609
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Commit-Queue: Weidong Liu <weidongliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293805}