[iOS] Remove BaseDefaultBrowserPromoSchedulerSceneAgent
BaseDefaultBrowserPromoSchedulerSceneAgent was introduced for
contextual full screen promo experimentation. It was used to separate
what promo is displayed and what metrics are recorded from other logic
so that different promos can be displayed using same triggers.
However, as that experiment is over having base class is not needed
anymore especially that it introduces complicated code paths that are
difficult to understand.

Bug: b/333883315
Change-Id: I7c9298bd6b90e42ec164a343caa38eedb33f5e5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5478188
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Commit-Queue: Gayane Petrosyan <gayane@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293807}