[ios] Fix MagicStack UI Breakage when Toggling Feed off
This issue is result of the MagicStackCollectionView pulling in the
safe area insets when it never is within the safe area. For some reason,
when toggling the feed off, the MagicStackCollectionView receives
safe area insets, even though its final position is well outside of
the safe area. By setting UIContentInsetsReferenceNone to the
UICollectionViewCompositionalLayoutConfiguration, the items are not
inset to safe area insets.

Fixed: 333745588
Change-Id: I4e60fe6d555de2a7a953df96559268f0f50d5a61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5474384
Reviewed-by: Scott Yoder <scottyoder@google.com>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293809}