CBD Timeframe Experiment: Remove "older than 30d" enum value
Remove the unnecessary "older than 30d" value from the WebUI TimeRange
enum for the experiment. Like the 15min value, the 30d value exists
for other platforms in the enum on the CPP side, but it's not relevant
for Desktop. For this reason the 30d value should not be in the WebUI
enum of the experiment - same as the 15min value is not in the regular
WebUI enum outside the experiment.

Bug: 40283307
Change-Id: Idd9cf94baec386d9dfd883d331879b4382b1697a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5494256
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Commit-Queue: Rainhard Findling <rainhard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293826}