WebUI: Update lit_migration.py script to handle bindings expressions.
The additional logic handles the most common cases of bindings. It is
not perfect, but it eliminates a lot of manual updates in practice.

Also fixing the regular expression handling event handler expressions
to also work when the event has a dash in the name.

Bug: 40943652
Change-Id: Ic72191ad43197f1d9268783601373e079d37a03c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5494687
Auto-Submit: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293827}