[Benefits] Check flags prior to loading credit card benefits
The PayDM Refresh method now has updated logic for loading credit
card benefits. In addition to checking the kAutofillPaymentCardBenefits
preference, it verifies if the kAutofillEnableCardBenefitsSync flag
is enabled. Benefits are loaded during refresh only if both conditions
are met: the sync flag and the preference are set to true.

Bug: 335253947
Change-Id: I14c3d544430b2d804d1eca6975d0daed8cf7e06e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5460449
Reviewed-by: Olivia Saul <jsaul@google.com>
Reviewed-by: Alexander Tekle <alexandertekle@google.com>
Commit-Queue: Viplav Kadam <viplavkadam@google.com>
Cr-Commit-Position: refs/heads/main@{#1293888}