Disable unsafe-buffers warnings when using precompiled headers
Precompiled headers are a serialized AST which needs to include the
pragmas used to turn those warnings on and off for various files and
expressions.

However since those pragmas are not serialized, the warnings fire in
places they should not, and break the build.

Precompiled headers are currently enabled only when remote compilation
is disabled, so we don't see these issues on our bots or on most
contributors' machines.

The `clang unsafe_buffer_usage` pragmas really should be serialized
and https://github.com/llvm/llvm-project/issues/90501 tracks that.

However the pragmas from our own plugin can not be de/serialized, as
there's no way to plug into the serialization path (that I can see).

R=hans

Bug: 41497066
Change-Id: If2c3211990646a4aebebcc658d383e9d726b679b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5497918
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293949}