[Extensions c2s] Update requests in Extensions UI only for current tab
Add and remove extension's site access requests from extensions toolbar
and extensions menu iff the request is for the current tab. For this, we
need to pass the tab id to the OnSiteAccessRequest[Added|Removed] observers.

Add a test that verifies extensions menu only shows requests for the
current tab. There is already an existent test for the extensions
toolbar, since behavior was already supported.

Bug: 330588494
Change-Id: I14fe0e29c0c6b37fbc8b7c80450e5a8d70f32001
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5493858
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293967}