Disable other touch variant of flaky test DetachFromFullsizeWindow
This test was observed to be flaky under the following conditions:

1. Tested against Ash M123
2. When using touch input

This is expected because there is a known issue with tests using touch
input that was only fixed in M124 (see http://crbug/323898910).
Previously only one of this test's touch variants was disabled. This
commit disables the other touch variant and moves both of them to the
correct section.

Bug: 337801917
Change-Id: I8ec02b4930c6fd1dd630305dac1444de98d33984
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5497383
Commit-Queue: Alex Yang <aycyang@chromium.org>
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293971}