[Site search policy] Surface featured SEs on @ state
Featured search engines set by policy are shown above the starter pack
suggestions when the user types "@" in the Omnibox.

The suggestion is very similar to the latest update of the starter pack
UI, including change to keyword mode when the suggestion is selected.

Screenshots: http://b/309458788#comment5

A follow-up will add the site favicon as specified in go/cbe-sitesearch.
The current icon is only a placeholder.

For additional protection, the change is protected by a new flag
`show-featured-enterprise-site-search`, so eventual regressions don't
block the policy roll out.

Bug: 309458788
Change-Id: I814c19aeec40c3c8c76bde9856116649f7d2568e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5443634
Reviewed-by: Angela Yoeurng <yoangela@chromium.org>
Commit-Queue: Fabio Tirelo <ftirelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293994}