[Views AX] Move FailFast to AXTreeManager
In preparation of moving BrowserAccessibility and
BrowserAccessibilityManager to the ui/ layer, I am working to move
anything that can be done ahead of time.

This change moves FailFast to AXTreeManager, allowing the
follow-up change to move the usage from BrowserAccessibilityManager
to AXPlatformTreeManager. In this, we can get rid of another TODO,
since this logic has been moved to the ui/ layer.

Bug: 40672441
Change-Id: Iefcf1fd8e5fa1df9fd84a9cec694f64bffaa3068
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5492381
Commit-Queue: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1294009}