[views-ax] Expose the right accessible name for SavedDeskItemView
SavedDeskItemView had one accessible name exposed in
GetAccessibleNodeData and a different one exposed through the
SetAccessibleName setter. I don't think this was on purpose. This CL
unifies both ways of setting the accessible name to the value that was
already exposed to the platform APIs -- the one in
GetAccessibleNodeData.

Essentially, the accessible name set through SetAccessibleName was the
visible name, and the one set in GetAccessibleNodeData was a string
composed of the desk template type and the template name. We chose the
latter, since the former was only used internally for testing and
never exposed to platform APIs.

This CL doesn't introduce a change in behavior. It's part of the
ViewsAX project. [1]

[1] https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417
Change-Id: I7f7dfcb12870375ce10fb422f895285c437c2a28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5499084
Auto-Submit: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1294091}